# swagger.json
{"openapi":"3.0.1","info":{"title":"APS Chat App API","description":"API documentation for APS Chat Application - A real-time chat application built with Spring Boot, WebSockets, and MongoDB","contact":{"name":"APS Team","email":"support@aps.com"},"license":{"name":"MIT License","url":"https://opensource.org/licenses/MIT"},"version":"1.0.0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"security":[{"JWT":[]}],"tags":[{"name":"Authentication","description":"Authentication endpoints for user login and token generation"},{"name":"User Management","description":"User registration and management endpoints"},{"name":"Messages","description":"Message management endpoints for chat functionality"}],"paths":{"/api/users/register":{"post":{"tags":["User Management"],"summary":"Register new user","description":"Register a new user in the system","operationId":"registerUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRegistrationRequest"}}},"required":true},"responses":{"201":{"description":"User registered successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}},"409":{"description":"Username already exists","content":{"*/*":{"schema":{"type":"object"}}}},"400":{"description":"Invalid user data or username already exists","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/messages":{"get":{"tags":["Messages"],"summary":"Get all messages","description":"Retrieve all chat messages with optional pagination","operationId":"getAllMessages","parameters":[{"name":"page","in":"query","description":"Page number (0-based)","required":false,"schema":{"type":"integer","format":"int32","default":0},"example":0},{"name":"size","in":"query","description":"Page size","required":false,"schema":{"type":"integer","format":"int32","default":20},"example":20},{"name":"sortDirection","in":"query","description":"Sort direction","required":false,"schema":{"type":"string","default":"DESC"},"example":"DESC"}],"responses":{"401":{"description":"Unauthorized - Invalid or missing JWT token","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageMessage"}}}},"200":{"description":"Messages retrieved successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Message"}}}}},"security":[{"JWT":[]}]},"post":{"tags":["Messages"],"summary":"Create new message","description":"Create a new chat message","operationId":"createMessage","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Message"}}},"required":true},"responses":{"201":{"description":"Message created successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Message"}}}},"400":{"description":"Invalid message data","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Message"}}}},"401":{"description":"Unauthorized - Invalid or missing JWT token","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Message"}}}}},"security":[{"JWT":[]}]}},"/api/auth/login":{"post":{"tags":["Authentication"],"summary":"User Login","description":"Authenticate user with username and password to receive JWT token","operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthRequest"}}},"required":true},"responses":{"200":{"description":"Authentication successful","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthResponse"}}}},"401":{"description":"Invalid username or password","content":{"*/*":{"schema":{"type":"object"}}}},"400":{"description":"Invalid request format","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/users":{"get":{"tags":["User Management"],"summary":"Get all users","description":"Retrieve all registered users (Admin only)","operationId":"getAllUsers","responses":{"401":{"description":"Unauthorized - Invalid or missing JWT token","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserResponse"}}}}},"403":{"description":"Forbidden - Admin access required","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserResponse"}}}}},"200":{"description":"Users retrieved successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}}},"security":[{"JWT":[]}]}},"/api/messages/{id}":{"get":{"tags":["Messages"],"summary":"Get message by ID","description":"Retrieve a specific message by its ID","operationId":"getMessageById","parameters":[{"name":"id","in":"path","description":"Message ID","required":true,"schema":{"type":"string"}}],"responses":{"401":{"description":"Unauthorized - Invalid or missing JWT token","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Message"}}}},"200":{"description":"Message found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Message"}}}},"404":{"description":"Message not found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Message"}}}}},"security":[{"JWT":[]}]},"delete":{"tags":["Messages"],"summary":"Delete message","description":"Delete a message by its ID","operationId":"deleteMessage","parameters":[{"name":"id","in":"path","description":"Message ID","required":true,"schema":{"type":"string"}}],"responses":{"401":{"description":"Unauthorized - Invalid or missing JWT token"},"404":{"description":"Message not found"},"204":{"description":"Message deleted successfully"}},"security":[{"JWT":[]}]}},"/api/messages/author/{autor}":{"get":{"tags":["Messages"],"summary":"Get messages by author","description":"Retrieve all messages from a specific author","operationId":"getMessagesByAuthor","parameters":[],"responses":{"401":{"description":"Unauthorized - Invalid or missing JWT token","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Message"}}}}},"200":{"description":"Messages retrieved successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Message"}}}}},"security":[{"JWT":[]}]}}},"components":{"schemas":{"UserRegistrationRequest":{"required":["password","username"],"type":"object","properties":{"username":{"type":"string","description":"Username for the new account","example":"john_doe"},"password":{"type":"string","description":"Password for the new account","example":"password123"}},"description":"User registration request"},"UserResponse":{"type":"object","properties":{"id":{"type":"string","description":"User ID","example":"64a1b2c3d4e5f6789abcdef0"},"username":{"type":"string","description":"Username","example":"john_doe"},"roles":{"uniqueItems":true,"type":"array","description":"User roles","example":["USER"],"items":{"type":"string","description":"User roles","example":"[\"USER\"]"}}},"description":"User information response"},"Message":{"type":"object","properties":{"id":{"type":"string","description":"Unique identifier for the message","example":"64a1b2c3d4e5f6789abcdef0"},"author":{"type":"string","description":"Message author username","example":"john_doe"},"content":{"type":"string","description":"Message content","example":"Hello, this is a test message!"},"timestamp":{"type":"string","description":"Message timestamp","format":"date-time"}},"description":"Chat message entity"},"AuthRequest":{"required":["password","username"],"type":"object","properties":{"username":{"type":"string","description":"Username for authentication","example":"john_doe"},"password":{"type":"string","description":"Password for authentication","example":"password123"}},"description":"Authentication request containing user credentials"},"AuthResponse":{"type":"object","properties":{"token":{"type":"string","description":"JWT token for authenticated requests","example":"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."}},"description":"Authentication response containing JWT token"},"PageMessage":{"type":"object","properties":{"totalElements":{"type":"integer","format":"int64"},"totalPages":{"type":"integer","format":"int32"},"first":{"type":"boolean"},"last":{"type":"boolean"},"size":{"type":"integer","format":"int32"},"content":{"type":"array","items":{"$ref":"#/components/schemas/Message"}},"number":{"type":"integer","format":"int32"},"sort":{"$ref":"#/components/schemas/SortObject"},"numberOfElements":{"type":"integer","format":"int32"},"pageable":{"$ref":"#/components/schemas/PageableObject"},"empty":{"type":"boolean"}}},"PageableObject":{"type":"object","properties":{"offset":{"type":"integer","format":"int64"},"sort":{"$ref":"#/components/schemas/SortObject"},"paged":{"type":"boolean"},"unpaged":{"type":"boolean"},"pageNumber":{"type":"integer","format":"int32"},"pageSize":{"type":"integer","format":"int32"}}},"SortObject":{"type":"object","properties":{"empty":{"type":"boolean"},"unsorted":{"type":"boolean"},"sorted":{"type":"boolean"}}}},"securitySchemes":{"JWT":{"type":"http","description":"JWT token for authentication. Format: 'Bearer <token>'","scheme":"bearer","bearerFormat":"JWT"}}}}